+++ tools/scripts-utils/src/path.js
@@ -27,7 +27,7 @@ function resolveBin(modName, { executable = modName, cwd = process.cwd() } = {})
 		const modPkgPath = require.resolve(`${modName}/package.json`);
 		const modPkgDir = path.dirname(modPkgPath);
 		const mod = JSON.parse(fs.readFileSync(modPkgPath));
-		const { bin } = mod.bin;
+		const bin = mod.bin;
 		const binPath = typeof bin === 'string' ? bin : bin[executable];
 		const fullPathToBin = path.join(modPkgDir, binPath);
 		if (fullPathToBin === systemCommandPath) {
